# sql: Create ExecutorContext to thread testing flags