--- a/sql/logic_test.go
+++ b/sql/logic_test.go
@@ -41,7 +41,6 @@ import (
 	"github.com/cockroachdb/cockroach/config"
 	"github.com/cockroachdb/cockroach/security"
 	"github.com/cockroachdb/cockroach/server"
-	csql "github.com/cockroachdb/cockroach/sql"
 	"github.com/cockroachdb/cockroach/testutils"
 	"github.com/cockroachdb/cockroach/testutils/sqlutils"
 	"github.com/cockroachdb/cockroach/util/leaktest"
@@ -244,6 +243,8 @@ func (t *logicTest) setup() {
 	// MySQL or Postgres instance.
 	ctx := server.NewTestContext()
 	ctx.MaxOffset = logicMaxOffset
+	ctx.TestingMocker.ExecutorTestingMocker.WaitForGossipUpdate = true
+	ctx.TestingMocker.ExecutorTestingMocker.CheckStmtStringChange = true
 	t.srv = setupTestServerWithContext(t.T, ctx)
 
 	// db may change over the lifetime of this function, with intermediate
@@ -693,7 +694,6 @@ func (t *logicTest) traceStop() {
 
 func TestLogic(t *testing.T) {
 	defer leaktest.AfterTest(t)()
-	defer csql.TestingWaitForGossipUpdate()()
 
 	// TODO(marc): splitting ranges at table boundaries causes
 	// a blocked task and won't drain. Investigate and fix.