@tamird 2016-03-08
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 5 of 5 files at r1.
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[sql/executor.go, line 172 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5009#-KCNP3MfhgP1q-cni7Fr:-KCNP3MgOz-pxZFKSRcF:941498707) ([raw file](https://github.com/cockroachdb/cockroach/blob/22dd7f461ff669168cc27c608989555faca64cd5/sql/executor.go#L172)):*
Instead of this comment, perhaps create a constructor? you could then unexport all these fields.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/5009#-:-KCNPGgSiJ9ia9_WMB_Q:1417088721)*
<!-- Sent from Reviewable.io -->