@nvanbenschoten 2016-03-08
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[sql/executor.go, line 172 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5009#-KCNP3MfhgP1q-cni7Fr:-KCNU-JwRcBAjekHf16o:-718593063) ([raw file](https://github.com/cockroachdb/cockroach/blob/22dd7f461ff669168cc27c608989555faca64cd5/sql/executor.go#L172)):*
I'm a little worried about creating a constructor if this Context struct continues to grow (which I assume it will). We also don't have any constructors for contexts anywhere else, so I'd like to stay consistent.

Also, the comment is directly stolen from `StoreContext`.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/5009)*
<!-- Sent from Reviewable.io -->